Allow creator to be empty for ApplicationEvent
We must update the status on an application programmatically via creating an ApplicationEvent. Since we would like to be able to set the status of an application to the default status for the program if a program has one, we need to be able to add an ApplicationEvent without a particular creator, as it won't be an admin actually changing the status of the application. This change allows the creator to be an Optional.empty() to allow for this. Note that we should ONLY use this for programmatic events like this, and never for an event generated by actual user input.