Add confirmation on changing default status
This adds a confirmation before changing the default status for a given program. Ideally, we'd trigger another, nicer modal from the edit status modal, but it turns out this is rather difficult with our framework. Instead, this utilizes a simple(ish) snippet of javascript that runs on submission that allows the user to cancel the action if they don't actually want to change the default status.